@@ -41,42 +41,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "4.3.1",
-    "@angular/common": "4.3.1",
-    "@angular/compiler": "4.3.1",
-    "@angular/core": "4.3.1",
-    "@angular/flex-layout": "2.0.0-beta.7",
-    "@angular/forms": "4.3.1",
-    "@angular/http": "4.3.1",
-    "@angular/material": "2.0.0-beta.7",
-    "@angular/platform-browser": "4.3.1",
-    "@angular/platform-browser-dynamic": "4.3.1",
-    "@angular/router": "4.3.1",
-    "@ng2-dynamic-forms/core": "1.4.23",
-    "@ng2-dynamic-forms/ui-bootstrap": "1.4.23",
-    "@ng2-dynamic-forms/ui-material": "1.4.23",
-    "@ngrx/effects": "4.0.1",
-    "@ngrx/router-store": "4.0.0",
-    "@ngrx/store": "4.0.0",
+    "@angular/animations": "4.4.1",
+    "@angular/cdk": "2.0.0-beta.10",
+    "@angular/common": "4.4.1",
+    "@angular/compiler": "4.4.1",
+    "@angular/core": "4.4.1",
+    "@angular/flex-layout": "2.0.0-beta.9",
+    "@angular/forms": "4.4.1",
+    "@angular/http": "4.4.1",
+    "@angular/material": "2.0.0-beta.10",
+    "@angular/platform-browser": "4.4.1",
+    "@angular/platform-browser-dynamic": "4.4.1",
+    "@angular/router": "4.4.1",
+    "@ng2-dynamic-forms/core": "1.4.31",
+    "@ng2-dynamic-forms/ui-bootstrap": "1.4.31",
+    "@ng2-dynamic-forms/ui-material": "1.4.31",
+    "@ngrx/effects": "4.0.5",
+    "@ngrx/router-store": "4.0.4",
+    "@ngrx/store": "4.0.3",
     "@ngrx/store-devtools": "4.0.0",
-    "@ngx-translate/core": "7.1.0",
-    "@ngx-translate/http-loader": "0.1.0",
+    "@ngx-translate/core": "8.0.0",
+    "@ngx-translate/http-loader": "2.0.0",
     "@swimlane/ngx-datatable": "10.0.5",
     "angular2-text-mask": "8.0.3",
-    "core-js": "2.4.1",
+    "core-js": "2.5.1",
     "hammerjs": "2.0.8",
     "lodash": "4.17.4",
-    "rxjs": "5.4.2",
-    "zone.js": "0.8.14"
+    "rxjs": "5.4.3",
+    "zone.js": "0.8.17"
   },
   "devDependencies": {
-    "@angular/cli": "1.3.0",
-    "@angular/compiler-cli": "4.3.1",
-    "@angular/language-service": "4.3.1",
+    "@angular/cli": "1.4.2",
+    "@angular/compiler-cli": "4.4.1",
+    "@angular/language-service": "4.4.1",
     "@types/jasmine": "2.5.53",
     "@types/jasminewd2": "2.0.2",
-    "@types/node": "8.0.20",
-    "codelyzer": "3.1.2",
+    "@types/node": "8.0.28",
+    "codelyzer": "3.2.0",
     "jasmine-core": "2.7.0",
     "jasmine-spec-reporter": "4.2.0",
     "karma": "1.7.0",
@@ -88,8 +89,8 @@
     "karma-mocha-reporter": "2.2.3",
     "protractor": "5.1.2",
     "ts-node": "3.3.0",
-    "tslint": "5.5.0",
+    "tslint": "5.7.0",
     "typedoc": "0.8.0",
-    "typescript": "2.4.2"
+    "typescript": "2.5.2"
   }
 }
