@@ -1,6 +1,7 @@
 import { routerReducer } from '@ngrx/router-store';
 import { ActionReducer } from '@ngrx/store';
 import { ActionReducerMap } from '@ngrx/store';
+import { MetaReducer } from '@ngrx/store';
 import { combineReducers } from '@ngrx/store';
 
 import { AUTH_LOGOUT } from '../../../modules/auth';
@@ -80,7 +81,7 @@ export function clear(reducer: ActionReducer<any>): ActionReducer<any> {
 /**
  * https://github.com/ngrx/platform
  */
-export const metaReducers: Array<ActionReducer<any>> = [
+export const metaReducers: Array<MetaReducer<any>> = [
   storage,
   clear,
 ];
