@@ -1,9 +1,9 @@
-import { Http } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 
 /**
  * https://github.com/ngx-translate/http-loader
  */
-export function TranslateLoaderFactory(http: Http): TranslateHttpLoader {
+export function TranslateLoaderFactory(http: HttpClient): TranslateHttpLoader {
   return new TranslateHttpLoader(http, './assets/i18ns/', '.json');
 }
